fix(draw): make _equipment_width_slot real int
The result of a division in Python is a float. The type of
equipment.type.width is the rack_width defined type which asserts its
value is a part of a whole one, thus making its inverted value an whole
number which can be safely converted into an integer without rounding.

+++ racksdb/drawers/infrastructure.py
@@ -214,7 +214,7 @@ class InfrastructureDrawer(Drawer):
 
     def _equipment_width_slot(self, equipment) -> int:
         """Return the slot of the equipment in the rack width."""
-        return (equipment.slot - equipment._first.slot) % (1 / equipment.type.width)
+        return (equipment.slot - equipment._first.slot) % int(1 / equipment.type.width)
 
     def _equipment_height_slot(self, equipment) -> int:
         """Return the slot of the equipment (ie. rack unit slot) in the rack height."""
